+++ src/pages/quiz/Wrong.tsx
@@ -1,26 +1,34 @@
 import Bg2 from '/src/assets/images/bg/bg2.png';
 import Avatar5 from '/src/assets/images/avatar/5.png';
-import Finish2 from './Finish2';
 import WhiteBox from '@/component/chatbox/WhiteBox';
-export default function Correct() {
-  // WhiteBox의 대사들
+interface WrongProps {
+  onRetry: () => void; // Define a prop for the retry function
+}
+
+const Wrong: React.FC<WrongProps> = ({ onRetry }) => {
   const dialogues = [
     <>
-      <span>정말 고맙네! 우리 연구생이 </span>
+      <span>그 곳이 아닌 거 같네...</span>
       <br />
-      <span>지관에 있었구만!</span>
+      <span>우리 지혜로운 연구생을</span>
+      <br />
+      <span>빨리 찾아주게나!</span>
     </>,
   ];
 
   return (
     <div className="flex justify-center w-full h-full bg-[#793A1C] relative">
-      <div className="absolute w-full h-full z-50">
-        <Finish2 />
-      </div>
-      <div className="flex justify-center w-full max-w-[500px] absolute bottom-[250px]">
+      <div className="w-full max-w-[500px] absolute bottom-[250px]">
         <img src={Bg2} />
       </div>
-      <div className="w-full p-8 max-w-[500px] absolute bottom-0 h-[250px] z-30 bg-[#661AAF]"></div>
+      <div className="w-full p-8 max-w-[500px] absolute bottom-0 h-[250px] z-30 bg-[#661AAF]">
+        <div
+          className={`flex items-center justify-center text-[1.2rem] p-5 text-white w-full text-center rounded-[15px] bg-[#0000007A]`}
+          onClick={onRetry}
+        >
+          다시
+        </div>
+      </div>
       <div
         className={`w-full max-w-[500px] absolute bottom-[250px] z-20 transition-transform duration-[2500ms]`}
       >
@@ -28,7 +36,7 @@ export default function Correct() {
           <img src={Avatar5} />
         </div>
       </div>
-      <div className="w-full h-[50%] overflow-y-scroll absolute top-[25%] p-4 max-w-[500px] space-y-4">
+      <div className="w-full h-[50%] absolute top-[25%] p-4 max-w-[500px] space-y-4">
         {dialogues.map((text, index) => (
           <div
             key={index}
@@ -40,4 +48,6 @@ export default function Correct() {
       </div>
     </div>
   );
-}
+};
+
+export default Wrong;
