+++ src/pages/quiz1/Finish2.tsx
