+++ src/pages/quiz1/Wrong.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import Bg2 from '/src/assets/images/bg/bg2.png';
 import Avatar5 from '/src/assets/images/avatar/5.png';
 import WhiteBox from '@/component/chatbox/WhiteBox';
